CarbonLess_backend/requirements.txt
fastapi[standard]
tortoise-orm==0.21.3
httpx==0.27.0
python-dotenv==1.0.0
asyncpg==0.30.0
aerich==0.9.2
pytest==8.4.2
pytest-asyncio==1.2.0
PyJWT==2.10.1
